# api/swagger.yaml
openapi: 3.0.1
info:
  title: Product Catalog Management
  description: "## TMF API Reference: TMF620 - Product Catalog Management\n\n### Release\
    \ : 19.0 - June 2019\n\nProduct Catalog API is one of Catalog Management API Family.\
    \ Product Catalog API goal is to provide a catalog of products. \n\n### Operations\n\
    Product Catalog API performs the following operations on the resources :\n- Retrieve\
    \ an entity or a collection of entities depending on filter criteria\n- Partial\
    \ update of an entity (including updating rules)\n- Create an entity (including\
    \ default values and creation rules)\n- Delete an entity\n- Manage notification\
    \ of events"
  version: 4.0.0
servers:
- url: https://serverRoot/tmf-api/productCatalogManagement/v4/
...
paths:
...
  /category:
    get:
      tags:
      - category
      summary: List or find Category objects
      description: This operation list or find Category entities
      operationId: listCategory
      parameters:
      - name: fields
        in: query
        description: Comma-separated properties to be provided in response
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: offset
        in: query
        description: Requested index for start of resources to be provided in response
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: limit
        in: query
        description: Requested number of resources to be provided in response
        required: false
        style: form
        explode: true
        schema:
          type: integer
      responses:
        200:
          description: Success
          headers:
            X-Total-Count:
              description: Total number of items matching criteria
              style: simple
              explode: false
              schema:
                type: integer
            X-Result-Count:
              description: Actual number of items returned in the response body
              style: simple
              explode: false
              schema:
                type: integer
          content:
            application/json;charset=utf-8:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
                x-content-type: application/json;charset=utf-8
        400:
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        405:
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: Category
    post:
      tags:
      - category
      summary: Creates a Category
      description: This operation creates a Category entity.
      operationId: createCategory
      requestBody:
        description: The Category to be created
        content:
          application/json;charset=utf-8:
            schema:
              $ref: '#/components/schemas/Category_Create'
        required: true
      responses:
        201:
          description: Created
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Category'
        400:
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        405:
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-codegen-request-body-name: category
      x-swagger-router-controller: Category
  /category/{id}:
    get:
      tags:
      - category
      summary: Retrieves a Category by ID
      description: This operation retrieves a Category entity. Attribute selection
        is enabled for all first level attributes.
      operationId: retrieveCategory
      parameters:
      - name: id
        in: path
        description: Identifier of the Category
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: fields
        in: query
        description: Comma-separated properties to provide in response
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: Success
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Category'
        400:
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        405:
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: Category
    delete:
      tags:
      - category
      summary: Deletes a Category
      description: This operation deletes a Category entity.
      operationId: deleteCategory
      parameters:
      - name: id
        in: path
        description: Identifier of the Category
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        204:
          description: Deleted
          content: {}
        400:
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        405:
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: Category
    patch:
      tags:
      - category
      summary: Updates partially a Category
      description: This operation updates partially a Category entity.
      operationId: patchCategory
      parameters:
      - name: id
        in: path
        description: Identifier of the Category
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: The Category to be updated
        content:
          application/json;charset=utf-8:
            schema:
              $ref: '#/components/schemas/Category_Update'
        required: true
      responses:
        200:
          description: Updated
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Category'
        400:
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        405:
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-codegen-request-body-name: category
      x-swagger-router-controller: Category
...
  /exportJob:
    get:
      tags:
      - exportJob
      summary: List or find ExportJob objects
      description: This operation list or find ExportJob entities
      operationId: listExportJob
      parameters:
      - name: fields
        in: query
        description: Comma-separated properties to be provided in response
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: offset
        in: query
        description: Requested index for start of resources to be provided in response
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: limit
        in: query
        description: Requested number of resources to be provided in response
        required: false
        style: form
        explode: true
        schema:
          type: integer
      responses:
        200:
          description: Success
          headers:
            X-Total-Count:
              description: Total number of items matching criteria
              style: simple
              explode: false
              schema:
                type: integer
            X-Result-Count:
              description: Actual number of items returned in the response body
              style: simple
              explode: false
              schema:
                type: integer
          content:
            application/json;charset=utf-8:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExportJob'
                x-content-type: application/json;charset=utf-8
        400:
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        405:
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: ExportJob
    post:
      tags:
      - exportJob
      summary: Creates a ExportJob
      description: This operation creates a ExportJob entity.
      operationId: createExportJob
      requestBody:
        description: The ExportJob to be created
        content:
          application/json;charset=utf-8:
            schema:
              $ref: '#/components/schemas/ExportJob_Create'
        required: true
      responses:
        201:
          description: Created
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/ExportJob'
        400:
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        405:
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-codegen-request-body-name: exportJob
      x-swagger-router-controller: ExportJob
  /exportJob/{id}:
    get:
      tags:
      - exportJob
      summary: Retrieves a ExportJob by ID
      description: This operation retrieves a ExportJob entity. Attribute selection
        is enabled for all first level attributes.
      operationId: retrieveExportJob
      parameters:
      - name: id
        in: path
        description: Identifier of the ExportJob
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: fields
        in: query
        description: Comma-separated properties to provide in response
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: Success
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/ExportJob'
        400:
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        405:
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: ExportJob
    delete:
      tags:
      - exportJob
      summary: Deletes a ExportJob
      description: This operation deletes a ExportJob entity.
      operationId: deleteExportJob
      parameters:
      - name: id
        in: path
        description: Identifier of the ExportJob
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        204:
          description: Deleted
          content: {}
        400:
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        405:
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: ExportJob
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the category
        href:
          type: string
          description: Reference of the category
        description:
          type: string
          description: Description of the category
        isRoot:
          type: boolean
          description: If true, this Boolean indicates that the category is a root
            of categories
        lastUpdate:
          type: string
          description: Date and time of the last update
          format: date-time
        lifecycleStatus:
          type: string
          description: Used to indicate the current lifecycle status
        name:
          type: string
          description: Name of the category
        parentId:
          type: string
          description: Unique identifier of the parent category
        version:
          type: string
          description: Category version
        productOffering:
          type: array
          description: A product offering represents entities that are orderable from
            the provider of the catalog, this resource includes pricing information.
          items:
            $ref: '#/components/schemas/ProductOfferingRef'
        subCategory:
          type: array
          description: The category resource is used to group product offerings, service
            and resource candidates in logical containers. Categories can contain
            other (sub-)categories and/or product offerings.
          items:
            $ref: '#/components/schemas/CategoryRef'
        validFor:
          $ref: '#/components/schemas/TimePeriod'
        '@baseType':
          type: string
          description: When sub-classing, this defines the super-class
        '@schemaLocation':
          type: string
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
        '@type':
          type: string
          description: When sub-classing, this defines the sub-class entity name
      description: The category resource is used to group product offerings, service
        and resource candidates in logical containers. Categories can contain other
        categories and/or product offerings, resource or service candidates.
      example:
        subCategory:
        - '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': http://example.com/aeiou
          version: version
        - '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': http://example.com/aeiou
          version: version
        isRoot: true
        lifecycleStatus: lifecycleStatus
        productOffering:
        - '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': http://example.com/aeiou
        - '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': http://example.com/aeiou
        validFor:
          startDateTime: 2000-01-23T04:56:07.000+00:00
          endDateTime: 2000-01-23T04:56:07.000+00:00
        '@type': '@type'
        description: description
        version: version
        parentId: parentId
        '@baseType': '@baseType'
        lastUpdate: 2000-01-23T04:56:07.000+00:00
        name: name
        id: id
        href: href
        '@schemaLocation': http://example.com/aeiou
    Category_Create:
      required:
      - name
      type: object
      properties:
        description:
          type: string
          description: Description of the category
        isRoot:
          type: boolean
          description: If true, this Boolean indicates that the category is a root
            of categories
        lastUpdate:
          type: string
          description: Date and time of the last update
          format: date-time
        lifecycleStatus:
          type: string
          description: Used to indicate the current lifecycle status
        name:
          type: string
          description: Name of the category
        parentId:
          type: string
          description: Unique identifier of the parent category
        version:
          type: string
          description: Category version
        productOffering:
          type: array
          description: A product offering represents entities that are orderable from
            the provider of the catalog, this resource includes pricing information.
          items:
            $ref: '#/components/schemas/ProductOfferingRef'
        subCategory:
          type: array
          description: The category resource is used to group product offerings, service
            and resource candidates in logical containers. Categories can contain
            other (sub-)categories and/or product offerings.
          items:
            $ref: '#/components/schemas/CategoryRef'
        validFor:
          $ref: '#/components/schemas/TimePeriod'
        '@baseType':
          type: string
          description: When sub-classing, this defines the super-class
        '@schemaLocation':
          type: string
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
        '@type':
          type: string
          description: When sub-classing, this defines the sub-class entity name
      description: |-
        The category resource is used to group product offerings, service and resource candidates in logical containers. Categories can contain other categories and/or product offerings, resource or service candidates.
        Skipped properties: id,href
    Category_Update:
      type: object
      properties:
        description:
          type: string
          description: Description of the category
        isRoot:
          type: boolean
          description: If true, this Boolean indicates that the category is a root
            of categories
        lifecycleStatus:
          type: string
          description: Used to indicate the current lifecycle status
        name:
          type: string
          description: Name of the category
        parentId:
          type: string
          description: Unique identifier of the parent category
        version:
          type: string
          description: Category version
        productOffering:
          type: array
          description: A product offering represents entities that are orderable from
            the provider of the catalog, this resource includes pricing information.
          items:
            $ref: '#/components/schemas/ProductOfferingRef'
        subCategory:
          type: array
          description: The category resource is used to group product offerings, service
            and resource candidates in logical containers. Categories can contain
            other (sub-)categories and/or product offerings.
          items:
            $ref: '#/components/schemas/CategoryRef'
        validFor:
          $ref: '#/components/schemas/TimePeriod'
        '@baseType':
          type: string
          description: When sub-classing, this defines the super-class
        '@schemaLocation':
          type: string
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
        '@type':
          type: string
          description: When sub-classing, this defines the sub-class entity name
      description: |-
        The category resource is used to group product offerings, service and resource candidates in logical containers. Categories can contain other categories and/or product offerings, resource or service candidates.
        Skipped properties: id,href,lastUpdate
...
    ExportJob:
      type: object
      properties:
        id:
          type: string
          description: Identifier of the export job
        href:
          type: string
          description: Reference of the export job
        completionDate:
          type: string
          description: Data at which the job was completed
          format: date-time
        contentType:
          type: string
          description: The format of the exported data
        creationDate:
          type: string
          description: Date at which the job was created
          format: date-time
        errorLog:
          type: string
          description: Reason for failure
        path:
          type: string
          description: URL of the root resource acting as the source for streaming
            content to the file specified by the export job
        query:
          type: string
          description: Used to scope the exported data
        url:
          type: string
          description: URL of the file containing the data to be exported
        status:
          $ref: '#/components/schemas/JobStateType'
        '@baseType':
          type: string
          description: When sub-classing, this defines the super-class
        '@schemaLocation':
          type: string
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
        '@type':
          type: string
          description: When sub-classing, this defines the sub-class entity name
      description: Represents a task used to export resources to a file
      example:
        '@type': '@type'
        query: query
        errorLog: errorLog
        creationDate: 2000-01-23T04:56:07.000+00:00
        url: url
        path: path
        '@baseType': '@baseType'
        completionDate: 2000-01-23T04:56:07.000+00:00
        id: id
        href: href
        '@schemaLocation': http://example.com/aeiou
        contentType: contentType
        status: Not Started
...
    JobStateType:
      type: string
      description: Valid values for the state of a batch job (e.g. catalog import)
      enum:
      - Not Started
      - Running
      - Succeeded
      - Failed
...
    Error:
      required:
      - code
      - reason
      type: object
      properties:
        code:
          type: string
          description: Application relevant detail, defined in the API or a common
            list.
        reason:
          type: string
          description: Explanation of the reason for the error which can be shown
            to a client user.
        message:
          type: string
          description: More details and corrective actions related to the error which
            can be shown to a client user.
        status:
          type: string
          description: HTTP Error code extension
        referenceError:
          type: string
          description: URI of documentation describing the error.
          format: uri
        '@baseType':
          type: string
          description: When sub-classing, this defines the super-class.
        '@schemaLocation':
          type: string
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
        '@type':
          type: string
          description: When sub-classing, this defines the sub-class entity name.
      description: Used when an API throws an Error, typically with a HTTP error response-code
        (3xx, 4xx, 5xx)